+++ CHANGELOG.md
@@ -7,6 +7,7 @@ All notable changes to this project will be documented in this file.
 - Implement a new compliance module `Supply Limit Module` that prevents minting more tokens that the specified limit.
 - Implement a new compliance module `Time Transfers limits` that prevents holders from transfering more token than a specified limit in a given time frame.
 - Implement two new compliance modules `Time Exchange limits` and `Monthly Exchange limits` that limit exchanges transfers. These are used to authorized specific trusted exchanges to hold tokens but limited to a certain amount transfered for a given time frame.
+- Add `function name() external pure returns (string memory);` to `IModule`. Compliance modules now require a `string constant public name = "ConditionalTransferModule";` constant variable to be declared.
 
 ## [4.0.1]